+++ next/.env.bratiska-cli-build.dev
@@ -1,5 +1,5 @@
 NEXT_PUBLIC_DEPLOYMENT=dev
-STRAPI_URL=https://olo-strapi.dev.bratislava.sk
+NEXT_PUBLIC_STRAPI_URL=https://olo-strapi.dev.bratislava.sk
 
 NEXT_PUBLIC_MEILISEARCH_SEARCH_API_KEY=
 NEXT_PUBLIC_MEILISEARCH_HOST=https://olo-strapi-meilisearch.dev.bratislava.sk/
