+++ next/.env.bratiska-cli-build.prod
@@ -1,6 +1,6 @@
 # Rename to .env.production.local when doing production deployment
 NEXT_PUBLIC_DEPLOYMENT=prod
-STRAPI_URL=https://olo-strapi.bratislava.sk
+NEXT_PUBLIC_STRAPI_URL=https://olo-strapi.bratislava.sk
 
 # NEXT_PUBLIC_MEILISEARCH_SEARCH_API_KEY=
 # NEXT_PUBLIC_MEILISEARCH_HOST=https://olo-strapi-meilisearch.bratislava.sk/
