+++ next/.env.bratiska-cli-build.staging
@@ -1,5 +1,5 @@
 NEXT_PUBLIC_DEPLOYMENT=staging
-STRAPI_URL=https://olo-strapi.staging.bratislava.sk
+NEXT_PUBLIC_STRAPI_URL=https://olo-strapi.staging.bratislava.sk
 
 # NEXT_PUBLIC_MEILISEARCH_SEARCH_API_KEY=
 # NEXT_PUBLIC_MEILISEARCH_HOST=https://olo-strapi-meilisearch.staging.bratislava.sk/
