+++ next/.env.example
@@ -1,6 +1,7 @@
 # Copy this to .env.local
+NEXT_PUBLIC_DEPLOYMENT=dev
 
-STRAPI_URL=http://127.0.0.1:1337
+NEXT_PUBLIC_STRAPI_URL=http://127.0.0.1:1337
 STRAPI_REVALIDATE_SECRET_TOKEN=TODO
 
 NEXT_PUBLIC_MEILISEARCH_SEARCH_API_KEY=TODO
