+++ next/codegen.ts
@@ -5,7 +5,7 @@ const codegenConfig: CodegenConfig = {
   schema: 'http://127.0.0.1:1337/graphql',
   documents: './src/services/graphql/queries/**/*.{gql,graphql}',
   generates: {
-    './src/services/graphql/index.ts': {
+    './src/services/graphql/api.ts': {
       plugins: ['typescript', 'typescript-operations', 'typescript-graphql-request'],
       // config: {
       //   enumsAsTypes: true
