+++ next/package.json
@@ -13,11 +13,12 @@
     "lint:quiet": "yarn lint --quiet",
     "lint:fix": "yarn lint --fix",
     "prettier": "prettier --write .",
-    "gen": "graphql-codegen && prettier --write ./src/services/graphql/index.ts",
+    "gen": "graphql-codegen && prettier --write src/services/graphql/api.ts",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build"
   },
   "dependencies": {
+    "@graphql-codegen/cli": "^5.0.2",
     "classnames": "^2.5.1",
     "graphql": "^16.8.1",
     "graphql-request": "^6.1.0",
@@ -34,7 +35,6 @@
   },
   "devDependencies": {
     "@chromatic-com/storybook": "1.3.1",
-    "@graphql-codegen/cli": "^5.0.2",
     "@graphql-codegen/typescript": "^4.0.6",
     "@graphql-codegen/typescript-graphql-request": "^6.2.0",
     "@graphql-codegen/typescript-operations": "^4.2.0",
@@ -88,7 +88,7 @@
     "eslint-plugin-xss": "^0.1.12",
     "postcss": "^8",
     "prettier": "^3.2.4",
-    "prettier-plugin-tailwindcss": "^0.5.11",
+    "prettier-plugin-tailwindcss": "^0.5.14",
     "tailwindcss": "^3.4.3",
     "typescript": "^5"
   },
