+++ next/src/pages/_app.tsx
@@ -20,12 +20,12 @@ const MyApp = ({ Component, pageProps }: AppProps) => {
         }
       `}</style>
       <Head>
-        <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
-        <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
-        <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
-        <link rel="manifest" href="/site.webmanifest" />
-        <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#e46054" />
-        <meta name="msapplication-TileColor" content="#da532c" />
+        {/* <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" /> */}
+        {/* <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" /> */}
+        {/* <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" /> */}
+        {/* <link rel="manifest" href="/site.webmanifest" /> */}
+        {/* <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#e46054" /> */}
+        {/* <meta name="msapplication-TileColor" content="#da532c" /> */}
         <meta name="theme-color" content="#ffffff" />
         {/* look for CookieConsent component for 3rd party scripts you'd expect to find here */}
       </Head>
